app/requirements.txt
Flask-SQLAlchemy
redlock-py
twilio
celery
supervisor-stdout
geopy
geocoder
rx
hazelcast-python-client
python-consul